django-review/tu_wallstreet/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
pytz==2022.1
sqlparse==0.4.2
requests
coverage